hubspot_integration/pyproject.toml
[tool.poetry]
name = "hubspot_integration"
version = "0.1.0"
description = "The program is implemented for HubSpot integration to DB. Currently, it supports the integration of company data."
authors = ["My Bui <my.bui.fi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
cassandra-driver = "^3.25.0"
hubspot-api-client = "^5.0.1"
python-dotenv = "^0.20.0"
certifi = "^2022.5.18"
click = "^8.1.3"
geomet = "0.2.1.post1"
python-dateutil = "^2.8.2"
six = "^1.16.0"
urllib3 = "^1.26.9"
pydantic = "^1.9.1"
types-python-dateutil = "^2.8.17"
core = {path = "../core", develop = true}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
autoflake = "^1.4"
mypy = "^0.942"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
